change: invoke `disconnected` before `[target]Disconnected` (#23)
In this PR, we made the decision to invoke the `disconnected` function
before invoking the `[target]Disconnected` function. This is because,
our apps have a tendency to clean up and refer targets within the
`disconnected` function, and if we disconnect targets before the element
itself, we get a null/undefined exception error and have to do safety
checks manually.